+++ Tuneful/Views/OnboardingView.swift
@@ -194,7 +194,7 @@ struct Shortcuts: View {
                     KeyboardShortcuts.Recorder("Next track:", name: .nextTrack)
                     KeyboardShortcuts.Recorder("Next track:", name: .previousTrack)
                     KeyboardShortcuts.Recorder("Toggle mini player:", name: .showMiniPlayer)
-                    KeyboardShortcuts.Recorder("Switch between Spotify and Apple Music:", name: .changeMusicPlayer)
+                    KeyboardShortcuts.Recorder("Switch music player:", name: .changeMusicPlayer)
                 }
                 .frame(maxWidth: .infinity, alignment: .center)